backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
bcrypt==4.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==25.0.3
werkzeug==3.1.5